Adicionado arquivo de execução do Quicksight.
